third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/svg-local-feimage-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not use cross-origin images in SVG filters within hidden containers</title>
  <link rel="help" href="https://github.com/WICG/html-in-canvas">
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='../support/getPixel.js'></script>
  <script src="/wpt_internal/resources/canvas-draw-element/waitForCanvasPaint.js"></script>
</head>
<body>
  <svg width="0" height="0">
    <defs>
      <filter id="filter-same" x="0" y="0" width="1" height="1">
        <feImage href="https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/resources/green-100x100.png" result="img"/>
        <feComposite in="img" in2="SourceGraphic" operator="over"/>
      </filter>

      <filter id="filter-cross" x="0" y="0" width="1" height="1">
        <feImage href="https://{{hosts[alt][www]}}:{{ports[https][0]}}/wpt_internal/html/canvas/drawElementImage/resources/red-100x100.png" result="img"/>
        <feComposite in="img" in2="SourceGraphic" operator="over"/>
      </filter>

      <pattern id="pattern-same" patternUnits="userSpaceOnUse" x="0" y="0" width="100" height="100">
        <rect width="100" height="100" fill="transparent" filter="url(#filter-same)"/>
      </pattern>

      <pattern id="pattern-cross" patternUnits="userSpaceOnUse" x="0" y="0" width="100" height="100">
        <rect width="100" height="100" fill="transparent" filter="url(#filter-cross)"/>
      </pattern>
    </defs>
  </svg>

  <canvas id="canvas" width="100" height="200" layoutsubtree>
    <svg width="100" height="200">
      <rect x="0" y="0" width="100" height="100" fill="url(#pattern-same)"/>
      <rect x="0" y="100" width="100" height="100" fill="url(#pattern-cross)"/>
    </svg>
  </canvas>

  <script>
    window.onload = () => {
      promise_test(async function(t) {
        const preloadImage = (url) => new Promise((resolve, reject) => {
          const img = new Image();
          img.onload = () => resolve(img);
          img.onerror = () => reject(new Error(`Failed to load image: ${url}`));
          img.src = url;
        });

        const sameOriginImageEl = document.querySelector('#filter-same feImage');
        const crossOriginImageEl = document.querySelector('#filter-cross feImage');

        await Promise.all([
          preloadImage(sameOriginImageEl.getAttribute('href')),
          preloadImage(crossOriginImageEl.getAttribute('href'))
        ]);

        await waitForCanvasPaint(canvas);
        var ctx = canvas.getContext('2d');
        ctx.fillStyle = 'blue';
        ctx.fillRect(0, 0, 100, 200);

        ctx.drawElementImage(document.querySelector('canvas > svg'), 0, 0);

        const imgData = ctx.getImageData(0, 0, canvas.width, canvas.height);

        let pixel = getPixel(imgData, 50, 50);
        assert_array_equals(pixel, [0, 255, 0, 255], "Same origin filter image in external pattern should be drawn");

        pixel = getPixel(imgData, 50, 150);
        assert_array_equals(pixel, [0, 0, 255, 255], "Cross origin filter image in external pattern should not be drawn");
      });
    }
  </script>
</body>
</html>
